requirements.txt
boto3
psycopg2-binary
tabulate
flask
python-dotenv
